requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
requests
datetime
dash==3.0.4
dash-bootstrap-components
pandas
plotly
plotly.express
numpy
scikit-fuzzy
scipy
networkx